+++ src/claude_compliance_sdk/__init__.py
@@ -33,6 +33,7 @@ from claude_compliance_sdk.exceptions import (
     BadRequestError,
     ComplianceClientError,
     ConflictError,
+    FileTooLargeError,
     InsufficientScopeError,
     InternalServerError,
     InvalidAPIKeyError,
@@ -42,6 +43,8 @@ from claude_compliance_sdk.exceptions import (
 )
 from claude_compliance_sdk.resources.activities import Activity
 from claude_compliance_sdk.resources.chats import Chat, ChatMessagesPage, Message
+from claude_compliance_sdk.resources.files import File
+from claude_compliance_sdk.resources.generated_files import GeneratedFile
 from claude_compliance_sdk.resources.organizations import Organization, User
 from claude_compliance_sdk.resources.project_documents import ProjectDocument
 from claude_compliance_sdk.resources.projects import Project, ProjectAttachment, ProjectDetail
@@ -64,6 +67,9 @@ __all__ = [
     "ComplianceClientError",
     "ConflictError",
     "CursorPage",
+    "File",
+    "FileTooLargeError",
+    "GeneratedFile",
     "InsufficientScopeError",
     "InternalServerError",
     "InvalidAPIKeyError",